requirements.txt
# ============================================================================
# LangGraph and LangChain Dependencies
# ============================================================================
langgraph>=0.2.0
langgraph-checkpoint>=1.0.0
langchain-openai>=0.2.0
langchain-core>=0.3.0

# ============================================================================
# Google API Dependencies
# ============================================================================
google-auth>=2.34.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.147.0

# ============================================================================
# HTTP and Web Scraping
# ============================================================================
requests>=2.32.0
beautifulsoup4>=4.12.0

# ============================================================================
# Environment and Configuration
# ============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0.0

# ============================================================================
# MCP (Model Context Protocol) Dependencies
# ============================================================================
mcp>=1.0.0

# ============================================================================
# Development Dependencies
# ============================================================================
# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Type Checking
mypy>=1.8.0
types-requests>=2.31.0

# Linting
ruff>=0.1.0

# ============================================================================
# Standard Library (no installation needed)
# ============================================================================
# base64 - for encoding Gmail messages
# logging - for workflow logging
# urllib - for URL parsing
# json - for JSON processing

